Create loading state for balances
