+++ app/utils/context.tsx
@@ -32,8 +32,9 @@ interface AppStateContextType {
   setSelectedChains: Dispatch<SetStateAction<ChainType[]>>;
   selectedWallet: WalletType | undefined;
   setSelectedWallet: Dispatch<SetStateAction<WalletType | undefined>>;
-  balanceList: WalletTokensData;
+  balanceList: WalletTokensData | undefined;
   refreshBalances: () => void;
+  isLoadingBalance: boolean
 }
 
 const AppStateContext = createContext<AppStateContextType | undefined>(
@@ -53,7 +54,7 @@ export const AppStateProvider = ({ children }: { children: ReactNode }) => {
     setIsWalletDrawerOpen((prevState) => !prevState);
   };
 
-  const { refreshBalances, balanceList } = useWalletTokens(walletsState!);
+  const { refreshBalances, balanceList, isLoadingBalance } = useWalletTokens(walletsState!);
 
   const checkAvailableWallets = (window: any) => {
     Object.keys(SUPPORTED_WALLETS).forEach((key) => {
@@ -259,6 +260,7 @@ export const AppStateProvider = ({ children }: { children: ReactNode }) => {
         setSelectedWallet,
         refreshBalances,
         balanceList,
+        isLoadingBalance
       }}
     >
       {children}
